requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
pydantic==2.5.0
pandas==2.1.3
plotly==5.18.0
requests==2.31.0
python-multipart==0.0.6
tensorflow==2.18.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.0